'''minor things'''
 * Function: remove the "auto-initialize" information callback
 * classificationEvaluator: score to minimize is 1.0 - accuracy
 * ProteinSequentialPredictor: rename sequencial into sequential

+++ include/lbcpp/Function/Evaluator.h
@@ -29,7 +29,7 @@ typedef ReferenceCountedObjectPtr<ScoreObject> ScoreObjectPtr;
 class CompositeScoreObject : public ScoreObject
 {
 public:
-  CompositeScoreObject() : scoreToMinimizeIndex(positiveIntegerType->getMissingValue().getInteger()) {}
+  CompositeScoreObject() : scoreToMinimizeIndex((size_t)positiveIntegerType->getMissingValue().getInteger()) {}
 
   virtual double getScoreToMinimize() const
     {return scoreToMinimizeIndex == (size_t)positiveIntegerType->getMissingValue().getInteger() ? 0.0 : scores[scoreToMinimizeIndex]->getScoreToMinimize();}
